migrate RenderDao to use mongodb java driver 3.0 model

+++ render-ws/src/main/java/org/janelia/render/service/RenderDataService.java
@@ -578,19 +578,19 @@ public class RenderDataService {
     @Path("project/{project}/stack/{stack}/z/{z}/box/{x},{y},{width},{height}/tile-count")
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public Integer getTileCount(@PathParam("owner") final String owner,
-                                @PathParam("project") final String project,
-                                @PathParam("stack") final String stack,
-                                @PathParam("x") final Double x,
-                                @PathParam("y") final Double y,
-                                @PathParam("z") final Double z,
-                                @PathParam("width") final Integer width,
-                                @PathParam("height") final Integer height) {
+    public Long getTileCount(@PathParam("owner") final String owner,
+                             @PathParam("project") final String project,
+                             @PathParam("stack") final String stack,
+                             @PathParam("x") final Double x,
+                             @PathParam("y") final Double y,
+                             @PathParam("z") final Double z,
+                             @PathParam("width") final Integer width,
+                             @PathParam("height") final Integer height) {
 
         LOG.info("getTileCount: entry, owner={}, project={}, stack={}, x={}, y={}, z={}, width={}, height={}",
                  owner, project, stack, x, y, z, width, height);
 
-        int count = 0;
+        long count = 0;
         try {
             final StackId stackId = new StackId(owner, project, stack);
             count = renderDao.getTileCount(stackId, x, y, z, width, height);
